refactor: load `routingConfig` into `window.guardian.routingConfig`
