+++ src/client/components/ClientState.tsx
@@ -1,15 +1,6 @@
 import React, { createContext, FunctionComponent } from 'react';
 
-import { ClientState } from '@/shared/model/ClientState';
-
-export const defaultClientState = {
-  clientHosts: { idapiBaseUrl: '', oauthBaseUrl: '' },
-  recaptchaConfig: { recaptchaSiteKey: '' },
-  sentryConfig: { build: '0', stage: 'DEV' as const, dsn: '' },
-  queryParams: {
-    returnUrl: 'https://theguardian.com',
-  },
-};
+import { ClientState, defaultClientState } from '@/shared/model/ClientState';
 
 export const ClientStateContext =
   createContext<ClientState>(defaultClientState);
