+++ src/shared/model/RoutingConfig.ts
@@ -1,0 +1,6 @@
+import { ClientState } from '@/shared/model/ClientState';
+
+export type RoutingConfig = {
+  clientState: ClientState;
+  location: string;
+};
